requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
boto3==1.24.36
botocore==1.27.36
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.4
django-allauth==0.51.0
django-ckeditor==6.4.0
django-cors-headers==3.13.0
django-filter==21.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-nested-admin==3.4.0
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg2==1.19.4
dynaconf==3.1.9
filelock==3.6.0
hvac==0.11.2
identify==2.5.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
Markdown==3.3.7
MarkupSafe==2.1.1
mysqlclient==2.1.0
nodeenv==1.6.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.49
Pillow==9.1.0
platformdirs==2.5.2
pre-commit==2.19.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
Serializer==0.2.1
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
validators==0.20.0
virtualenv==20.14.1
zipp==3.8.0
